Bug in Button Press
MakeCode Editor had an update that broke P0 (up button) in the extension. I'm not sure why the button press was working before  and now it isn't? The other buttons work so I assume that it is just the `<number>`. I checked other syntax that uses `DAL.MICROBIT_ID_IO_P0` and it seems this might be the issue.

I also bumped up the version number and updated the note in the README.md.